Change datatype of `i` and PIN-numbers to byte
PIN_LATCH, PIN_CLOCK, PIN_DATA and i won't go above 255, so a byte is sufficient to store those values. Don't waste too much RAM.

+++ SNESpaduino.h
@@ -20,7 +20,7 @@ class SNESpaduino
 		uint16_t getButtons(void);
 	private:
 		uint16_t state;
-		int i, PIN_LATCH, PIN_CLOCK, PIN_DATA;
+		byte i, PIN_LATCH, PIN_CLOCK, PIN_DATA;
 };
 
 #endif
